Avoid import cycle.
